+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 ### Changed
 
 - Adjust theme.
+- Enable feature based routes and cookie policy, privacy policy and terms pages in config.
 - Update to base version [`2.5.0`](https://github.com/slsfi/digital-edition-frontend-ng/releases/tag/2.5.0) from upstream, original repository.